* added a save() method to write configuration files
* fixed up the clone method to use the class methods to copy
directives and values

* added bits to remember which directives were found in each file.
I'm not sure what to use that for.  I'm nervous about writing values
back to files because later values in the objects would replace
that value in all files, even if they should stay the same (like
in a global file that should be overriden rather than replaced).

* added tests in save.t, but I still need to peek in the files to
make sure that the right things are in there.


git-svn-id: http://brian-d-foy.svn.sourceforge.net/svnroot/brian-d-foy/trunk@1218 d9f60a02-fc28-0410-9752-bc0d6c4f57a6